Move cmd/storage to impl/sql/engine
