+++ cmd/key-transparency-signer/backend.go
@@ -24,12 +24,12 @@ import (
 	"sync"
 	"time"
 
-	"github.com/google/key-transparency/cmd/storage"
 	"github.com/google/key-transparency/core/mutator/entry"
 	"github.com/google/key-transparency/core/signatures"
 	"github.com/google/key-transparency/core/signer"
 	"github.com/google/key-transparency/impl/etcd/queue"
 	"github.com/google/key-transparency/impl/sql/appender"
+	"github.com/google/key-transparency/impl/sql/engine"
 	"github.com/google/key-transparency/impl/sql/sqlhist"
 	"github.com/google/key-transparency/impl/transaction"
 
@@ -47,7 +47,7 @@ var (
 )
 
 func openDB() *sql.DB {
-	db, err := sql.Open(storage.DriverName, *serverDBPath)
+	db, err := sql.Open(engine.DriverName, *serverDBPath)
 	if err != nil {
 		log.Fatalf("sql.Open(): %v", err)
 	}
